svd2rust update: 0.30.0 - 0.36.0
There has been an issue[1] with `proc-macro2` which svd2rust depends on.

It is required to upgrade `svd2rust >= v0.33.5` but since we do the
upgrade now, lets just kick it up to `v0.36.0`.

This upgrade is quite seamless but requires some modification to the
`Config` struct when using `svd2rust` `generation`.

The update to the `atsamx7x` hal is a bit more complex, but can be done
with a view regex edits.

[1] https://github.com/rust-embedded/svd2rust/issues/863